WIP: color editor enhancements
