+++ src/App/Vivianne.Wpf/Views/FcePreviewView.xaml
@@ -80,9 +80,9 @@
         <Border Grid.Row="1" Grid.Column="1">
             <Viewport3D>
                 <Viewport3D.Camera>
-                    <PerspectiveCamera                        
-                    x:Name="ptcMain"
-                    UpDirection="0,0,1"
+                    <PerspectiveCamera
+                        x:Name="ptcMain"
+                        UpDirection="0,0,1"
                     LookDirection="40,-40,-40"
                     Position="-40,40,40"/>
                 </Viewport3D.Camera>
