+++ src/App/Vivianne.Wpf/Views/FcePreviewView.xaml.cs
@@ -20,18 +20,17 @@ public partial class FcePreviewView : UserControl
     Point from;
     private void Window_PreviewMouseMove(object sender, MouseEventArgs e)
     {
-             var till = e.GetPosition(sender as IInputElement);
-            double dx = till.X - from.X;
-            double dy = till.Y - from.Y;
-            from = till;
+        var till = e.GetPosition(sender as IInputElement);
+        double dx = till.X - from.X;
+        double dy = till.Y - from.Y;
+        from = till;
 
-            var distance = dx * dx + dy * dy;
-            if (distance <= 0d)  return;
-       if (e.MouseDevice.LeftButton is MouseButtonState.Pressed)
+        var distance = dx * dx + dy * dy;
+        if (distance <= 0d) return;
+        if (e.MouseDevice.LeftButton is MouseButtonState.Pressed)
         {
-
             var angle = (distance / ptcMain.FieldOfView) % 45d;
-            ptcMain.Rotate(new(dx*2, dy*3, 0d), angle);
+            ptcMain.Rotate(new(dx * 2, dy * 3, 0d), angle);
         }
     }
 }
